created Hyphens component
